+++ internal/adapters/http/pkg/app/config/init_test.go
@@ -23,7 +23,7 @@ func TestRebuild(t *testing.T) {
 	}
 
 	network := ""
-	for k := range build.GNetworks {
+	for k := range build.GetNetworks() {
 		network = k
 		break
 	}