+++ internal/adapters/http/pkg/app/errors.go
@@ -13,9 +13,11 @@ func (err *SAppError) Error() string {
 }
 
 var (
-	ErrRunning = &SAppError{"app running"}
-	ErrService = &SAppError{"service"}
-	ErrClose   = &SAppError{"close"}
-	ErrInitDB  = &SAppError{"init database"}
-	ErrExist   = &SAppError{"exist"}
+	ErrRunning     = &SAppError{"app running"}
+	ErrService     = &SAppError{"service"}
+	ErrClose       = &SAppError{"close"}
+	ErrInitDB      = &SAppError{"init database"}
+	ErrExist       = &SAppError{"exist"}
+	ErrInitConfig  = &SAppError{"init config"}
+	ErrSetNetworks = &SAppError{"set networks"}
 )