+++ internal/applications/messenger/pkg/app/init_app.go
@@ -7,18 +7,27 @@ import (
 
 	"github.com/number571/go-peer/pkg/types"
 	"github.com/number571/hidden-lake/internal/applications/messenger/pkg/app/config"
-	"github.com/number571/hidden-lake/internal/applications/messenger/pkg/settings"
+	hlm_settings "github.com/number571/hidden-lake/internal/applications/messenger/pkg/settings"
+	"github.com/number571/hidden-lake/internal/utils/build"
 	"github.com/number571/hidden-lake/internal/utils/flag"
+	std_logger "github.com/number571/hidden-lake/internal/utils/logger/std"
 )
 
 func InitApp(pArgs []string, pFlags flag.IFlags) (types.IRunner, error) {
 	inputPath := strings.TrimSuffix(pFlags.Get("-p").GetStringValue(pArgs), "/")
+	okLoaded, err := build.SetBuildByPath(inputPath)
+	if err != nil {
+		return nil, errors.Join(ErrSetNetworks, err)
+	}
 
-	cfgPath := filepath.Join(inputPath, settings.CPathYML)
+	cfgPath := filepath.Join(inputPath, hlm_settings.CPathYML)
 	cfg, err := config.InitConfig(cfgPath, nil)
 	if err != nil {
 		return nil, errors.Join(ErrInitConfig, err)
 	}
+
+	stdfLogger := std_logger.NewStdLogger(cfg.GetLogging(), std_logger.GetLogFunc())
+	build.LogLoadedBuildFiles(hlm_settings.GServiceName, stdfLogger, okLoaded)
 
 	return NewApp(cfg, inputPath), nil
 }