+++ internal/composite/pkg/app/config/init.go
@@ -52,7 +52,7 @@ func rebuildConfig(pCfg IConfig, pUseNetwork string) (IConfig, error) {
 	}
 
 	cfg := pCfg.(*SConfig)
-	network, ok := build.GNetworks[pUseNetwork]
+	network, ok := build.GetNetwork(pUseNetwork)
 	if !ok {
 		return nil, errors.Join(ErrRebuildConfig, ErrNetworkNotFound)
 	}