+++ internal/composite/pkg/app/init_app.go
@@ -7,8 +7,10 @@ import (
 
 	"github.com/number571/go-peer/pkg/types"
 	"github.com/number571/hidden-lake/internal/composite/pkg/app/config"
-	"github.com/number571/hidden-lake/internal/composite/pkg/settings"
+	hlc_settings "github.com/number571/hidden-lake/internal/composite/pkg/settings"
+	"github.com/number571/hidden-lake/internal/utils/build"
 	"github.com/number571/hidden-lake/internal/utils/flag"
+	std_logger "github.com/number571/hidden-lake/internal/utils/logger/std"
 
 	hla_tcp_app "github.com/number571/hidden-lake/internal/adapters/tcp/pkg/app"
 	hla_tcp_settings "github.com/number571/hidden-lake/internal/adapters/tcp/pkg/settings"
@@ -34,8 +36,12 @@ import (
 
 func InitApp(pArgs []string, pFlags flag.IFlags) (types.IRunner, error) {
 	inputPath := strings.TrimSuffix(pFlags.Get("-p").GetStringValue(pArgs), "/")
+	okLoaded, err := build.SetBuildByPath(inputPath)
+	if err != nil {
+		return nil, errors.Join(ErrSetNetworks, err)
+	}
 
-	cfgPath := filepath.Join(inputPath, settings.CPathYML)
+	cfgPath := filepath.Join(inputPath, hlc_settings.CPathYML)
 	cfg, err := config.InitConfig(cfgPath, nil, pFlags.Get("-n").GetStringValue(pArgs))
 	if err != nil {
 		return nil, errors.Join(ErrInitConfig, err)
@@ -45,6 +51,9 @@ func InitApp(pArgs []string, pFlags flag.IFlags) (types.IRunner, error) {
 	if err != nil {
 		return nil, errors.Join(ErrGetRunners, err)
 	}
+
+	stdfLogger := std_logger.NewStdLogger(cfg.GetLogging(), std_logger.GetLogFunc())
+	build.LogLoadedBuildFiles(hlc_settings.GServiceName, stdfLogger, okLoaded)
 
 	return NewApp(cfg, runners), nil
 }