+++ internal/service/internal/handler/network_request_test.go
@@ -373,7 +373,7 @@ func testNewPushNode(cfgPath, dbPath string) (anonymity.INode, context.CancelFun
 	)
 
 	node.HandleFunc(
-		build.GSettings.FProtoMask.FService,
+		build.GetSettings().FProtoMask.FService,
 		handler.RequestHandler(HandleServiceFunc(cfg, logger)),
 	)
 	node.GetMapPubKeys().SetPubKey(tgPrivKey1.GetPubKey())