+++ internal/service/pkg/app/config/init.go
@@ -37,7 +37,7 @@ func rebuildConfig(pCfg IConfig, pUseNetwork string) (IConfig, error) {
 	}
 
 	cfg := pCfg.(*SConfig)
-	network, ok := build.GNetworks[pUseNetwork]
+	network, ok := build.GetNetwork(pUseNetwork)
 	if !ok {
 		return nil, errors.Join(ErrRebuildConfig, ErrNetworkNotFound)
 	}
@@ -61,7 +61,7 @@ func rebuildConfig(pCfg IConfig, pUseNetwork string) (IConfig, error) {
 }
 
 func initConfig() *SConfig {
-	defaultNetwork := build.GNetworks[build.CDefaultNetwork]
+	defaultNetwork, _ := build.GetNetwork(build.CDefaultNetwork)
 	return &SConfig{
 		FSettings: &SConfigSettings{
 			FMessageSizeBytes: defaultNetwork.FMessageSizeBytes,