+++ internal/service/pkg/app/init_app.go
@@ -6,24 +6,33 @@ import (
 	"strings"
 
 	"github.com/number571/go-peer/pkg/types"
+	"github.com/number571/hidden-lake/internal/utils/build"
 	"github.com/number571/hidden-lake/internal/utils/flag"
+	std_logger "github.com/number571/hidden-lake/internal/utils/logger/std"
 	"github.com/number571/hidden-lake/internal/utils/privkey"
 
 	"github.com/number571/hidden-lake/internal/service/pkg/app/config"
-	pkg_settings "github.com/number571/hidden-lake/internal/service/pkg/settings"
+	hls_settings "github.com/number571/hidden-lake/internal/service/pkg/settings"
 )
 
 // initApp work with the raw data = read files, read args
 func InitApp(pArgs []string, pFlags flag.IFlags) (types.IRunner, error) {
 	inputPath := strings.TrimSuffix(pFlags.Get("-p").GetStringValue(pArgs), "/")
+	okLoaded, err := build.SetBuildByPath(inputPath)
+	if err != nil {
+		return nil, errors.Join(ErrSetNetworks, err)
+	}
 
-	cfgPath := filepath.Join(inputPath, pkg_settings.CPathYML)
+	cfgPath := filepath.Join(inputPath, hls_settings.CPathYML)
 	cfg, err := config.InitConfig(cfgPath, nil, pFlags.Get("-n").GetStringValue(pArgs))
 	if err != nil {
 		return nil, errors.Join(ErrInitConfig, err)
 	}
 
-	keyPath := filepath.Join(inputPath, pkg_settings.CPathKey)
+	stdfLogger := std_logger.NewStdLogger(cfg.GetLogging(), std_logger.GetLogFunc())
+	build.LogLoadedBuildFiles(hls_settings.GServiceName, stdfLogger, okLoaded)
+
+	keyPath := filepath.Join(inputPath, hls_settings.CPathKey)
 	privKey, err := privkey.GetPrivKey(keyPath)
 	if err != nil {
 		return nil, errors.Join(ErrGetPrivateKey, err)