+++ pkg/network/network.go
@@ -35,6 +35,7 @@ func NewHiddenLakeNode(
 	pRunnerAdapter adapters.IRunnerAdapter,
 	pHandlerF handler.IHandlerF,
 ) IHiddenLakeNode {
+	buildSettings := build.GetSettings()
 	adaptersSettings := pSettings.GetAdapterSettings()
 	return &sHiddenLakeNode{
 		anonymity.NewNode(
@@ -51,12 +52,12 @@ func NewHiddenLakeNode(
 						FSettings: adaptersSettings,
 						FParallel: pSettings.GetPowParallel(),
 					}),
-					FNetworkMask:  build.GSettings.FProtoMask.FNetwork,
+					FNetworkMask:  buildSettings.FProtoMask.FNetwork,
 					FQueuePeriod:  pSettings.GetQueuePeriod(),
 					FConsumersCap: pSettings.GetQBPConsumers(),
 					FQueuePoolCap: [2]uint64{
-						build.GSettings.FQueueBasedProblem.FMainPoolCap,
-						build.GSettings.FQueueBasedProblem.FRandPoolCap,
+						buildSettings.FQueueBasedProblem.FMainPoolCap,
+						buildSettings.FQueueBasedProblem.FRandPoolCap,
 					},
 				}),
 				func() client.IClient {
@@ -68,7 +69,7 @@ func NewHiddenLakeNode(
 				}(),
 			),
 		).HandleFunc(
-			build.GSettings.FProtoMask.FService,
+			buildSettings.FProtoMask.FService,
 			handler.RequestHandler(pHandlerF),
 		),
 	}
@@ -119,11 +120,12 @@ func (p *sHiddenLakeNode) SendRequest(
 	pPubKey asymmetric.IPubKey,
 	pRequest request.IRequest,
 ) error {
+	buildSettings := build.GetSettings()
 	err := p.fOriginNode.SendPayload(
 		pCtx,
 		pPubKey,
 		payload.NewPayload64(
-			uint64(build.GSettings.FProtoMask.FService),
+			uint64(buildSettings.FProtoMask.FService),
 			pRequest.ToBytes(),
 		),
 	)
@@ -138,11 +140,12 @@ func (p *sHiddenLakeNode) FetchRequest(
 	pPubKey asymmetric.IPubKey,
 	pRequest request.IRequest,
 ) (response.IResponse, error) {
+	buildSettings := build.GetSettings()
 	rspBytes, err := p.fOriginNode.FetchPayload(
 		pCtx,
 		pPubKey,
 		payload.NewPayload32(
-			build.GSettings.FProtoMask.FService,
+			buildSettings.FProtoMask.FService,
 			pRequest.ToBytes(),
 		),
 	)